requirements.txt
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
ffmpeg==1.4
filelock==3.16.1
Flask==3.0.3
fsspec==2024.9.0
gunicorn==23.0.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
packaging==24.1
pillow==10.4.0
pocketsphinx==5.0.3
proglog==0.1.10
pycparser==2.22
pydub==0.25.1
regex==2024.9.11
requests==2.32.3
setuptools==75.1.0
sounddevice==0.5.0
SpeechRecognition==3.10.4
sympy==1.13.3
tiktoken==0.7.0
tk==0.1.0
torch==2.4.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
